+++ src/cutouts_service/cutouts/__init__.py
@@ -6,9 +6,11 @@ from cutouts_service.cutouts.cutout import (
     Options,
 )
 from cutouts_service.cutouts.astropy_cutout import AstropyCutout
+from cutouts_service.cutouts.objstore_cutout import ObjStoreCutout
 
 __all__ = [
     "AstropyCutout",
+    "ObjStoreCutout",
     "ImageLikeHDU",
     "IOConfig",
     "CutoutConfig",