+++ src/cutouts_service/cutouts/astropy_cutout.py
@@ -26,7 +26,22 @@ logger = logging.getLogger(__name__)
 
 
 class AstropyCutout(Cutout):
-    """A cutout class encapsulating the cutout from a remote source using astropy"""
+    """A cutout class encapsulating the cutout from a remote source using astropy
+
+    Parameters
+    ----------
+    io_config : IOConfig
+        The config describing the IO details (eg. url)
+    cutout_config : CutoutConfig
+        The config describing the cutout details (eg. pointing)
+    options : Options
+        The extra options, currently contains only dry_run
+
+    Attributes
+    ----------
+    source_header: fits.Header
+        The header of the source file
+    """
 
     def __init__(
         self,
@@ -80,7 +95,7 @@ class AstropyCutout(Cutout):
 
     def _build_spatial_cutout(
         self, image_hdu: ImageLikeHDU
-    ) -> tuple[np.ndarray, fits.Header, list[slice]]:
+    ) -> tuple[np.ndarray, fits.Header, tuple[slice]]:
         """Generate a cutout of a fits file
 
         Parameters
@@ -90,7 +105,7 @@ class AstropyCutout(Cutout):
 
         Returns
         -------
-        tuple[np.ndarray, fits.Header, list[slice]]
+        tuple[np.ndarray, fits.Header, tuple[slice]]
             The data array, header, and slices generated for this cutout
 
         Raises
@@ -148,6 +163,11 @@ class AstropyCutout(Cutout):
     def create_cutout(self, overwrite: bool = False) -> Path:
         """Extract a sky cutout and write it to a FITS file.
 
+        Parameters
+        ----------
+        overwrite : bool
+            Allow overwriting the output file
+
         Returns
         -------
         Path
@@ -182,6 +202,10 @@ class AstropyCutout(Cutout):
             self.source_header = image_hdu.header
 
             self._compute_pixel_indices()
+            if not self.check_cutout_fit():
+                logger.warning(
+                    "The provided cutout configuration extends past the extents of the selected cube. The cutout extents will be clipped appropriately."
+                )
 
             if self.dry_run:
                 self._get_cube_details(image_hdu)