+++ src/cutouts_service/cutouts/cutout.py
@@ -30,6 +30,17 @@ ImageLikeHDU = fits.PrimaryHDU | fits.ImageHDU | fits.CompImageHDU
 
 @dataclass
 class IOConfig:
+    """The IO details
+    
+    Parameters
+    ----------
+    source : str | Path
+        The source path or url
+    output_path : str | Path
+        The path to for the output file
+    s3_endpoint_url : str | None, Optional
+        The s3 endpoint url, default is None
+    """
     source: str | Path
     output_path: str | Path
     s3_endpoint_url: str | None = None
@@ -37,6 +48,19 @@ class IOConfig:
 
 @dataclass
 class CutoutConfig:
+    """The cutout details
+
+    Parameters
+    ----------
+    ra : float
+        The Right ascension of the pointing in decimal degrees
+    dec : float
+        The declination of the pointing in decimal degrees
+    radius : float
+        The radius/extent of the cutout centred at the pointing
+    channel_range : tuple[int, ...] | tuple[None, ...], Optional
+        The inclusive channel range to cutout on the spectral axis
+    """
     ra: float
     dec: float
     radius: float
@@ -45,11 +69,39 @@ class CutoutConfig:
 
 @dataclass
 class Options:
+    """The options
+
+    Parameters
+    ----------
+    dry_run : bool
+        Run a dry-run instead to check input parameters
+    """
     dry_run: bool = False
 
 
 class Cutout(ABC):
-    """A general cutout class that needs a to be overwritten with a specific tool"""
+    """A general cutout class that needs a to be overwritten with a specific tool
+    
+    Parameters
+    ----------
+    io_config : IOConfig
+        The config describing the IO details (eg. url)
+    cutout_config : CutoutConfig
+        The config describing the cutout details (eg. pointing)
+    options : Options
+        The extra options, currently contains only dry_run
+
+    Attributes
+    ----------
+    source_header : dict[str, Any]
+        The header of the source file
+    fits_shape : tuple[int, ...]
+        The shape of the fits cube
+    pixel_indices : dict[str, int]
+        The pixel indices of the cutout
+    axis_types : tuple[str]
+        The types of the relevant axes
+    """
 
     def __init__(
         self,
@@ -71,6 +123,11 @@ class Cutout(ABC):
     def _set_header_shape(self, header: dict[str, Any]):
         """Get the shape of the fits file form the header
 
+        Parameters
+        ----------
+        header : dict[str, Any]
+            The header file
+
         Returns
         -------
         tuple[int, ...]
@@ -89,29 +146,14 @@ class Cutout(ABC):
 
     def _compute_pixel_indices(self):
         """Compute the array indices from the input celestial coordinates
-
-        Parameters
-        ----------
-        header : fits.Header
-            The header containing the WCS information needed to do the conversion
-        position : SkyCoord
-            The sky position of the center of the cutout
-        size : u.Quantity
-            The size of the intended cutout (assuming a square cutout)
-
-        Returns
-        -------
-        dict[str, int | list[str]]
-            Contains the pixel extents within the fits array as well as a list of axis types
         """
 
         header = self.source_header
-
         position = SkyCoord(
             ra=self.cutout_config.ra * u.deg, dec=self.cutout_config.dec * u.deg
         )
         size = 2 * self.cutout_config.radius * u.deg
-        wcs = WCS(header)
+        wcs = WCS(fits.Header(header))
         ra_dec_min = position.spherical_offsets_by(-size / 2, -size / 2)
         ra_dec_max = position.spherical_offsets_by(size / 2, size / 2)
         x0, y0 = wcs.celestial.world_to_pixel(ra_dec_min)
@@ -134,10 +176,41 @@ class Cutout(ABC):
 
     @abstractmethod
     def create_cutout(self, overwrite: bool = False):
-        """Extract a sky cutout and write it to a FITS file."""
+        """Extract a sky cutout and write it to a FITS file.
+        
+        Parameters
+        ----------
+        overwrite : bool
+            Allow overwriting the output file
+
+        Raises
+        ------
+        NotImplementedError
+            The method needs to be overwritten
+        """
         raise NotImplementedError(
             "This method must be overwritten with the specific cutout implementation"
         )
+
+    def check_cutout_fit(self) -> bool:
+        """Checks if the requested cutout fits in the given cube
+
+        Returns
+        -------
+        bool
+            True if the cutout fits, False if it doesn't.
+        """
+        shape = self.fits_shape[::-1]
+        chans = self.cutout_config.channel_range
+        cutout_indices = self.pixel_indices
+        if cutout_indices["xmin"] < 0 or cutout_indices["xmax"] > (shape[0] - 1):
+            return False
+        if cutout_indices["ymin"] < 0 or cutout_indices["ymax"] > (shape[1] - 1):
+            return False
+        if chans[0] is not None and chans[1] is not None:
+            if chans[0] < 0 or chans[1] > (shape[-1] - 1):
+                return False
+        return True
 
     def _get_cube_details(self):
         """Query and print key Cube details from header"""
@@ -188,13 +261,13 @@ class Cutout(ABC):
             )
 
     def build_cutout_header(
-        self, slices: list[slice], shape: tuple[int, ...], section_dtype: np.dtype
+        self, slices: tuple[slice], shape: tuple[int, ...], section_dtype: np.dtype
     ) -> fits.Header:
         """Return a FITS header adjusted for a cutout region.
 
         Parameters
         ----------
-        slices : list[slice]
+        slices : tuple[slice]
             The slices that were performed on the fits data
         shape : tuple[int, ...]
             The shape of the data